fix(native): move L111 graph rule to graph package (ADR-059)
